transflow/requirements.txt
fastapi
uvicorn
pymongo
redis
faststream
faststream[rabbit]
python-dotenv
